+++ src/Pim/Bundle/CatalogBundle/Model/AvailableAttributes.php
@@ -2,8 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Model;
 
-use Doctrine\Common\Collections\ArrayCollection;
-
 /**
  * Available attributes model
  *
@@ -14,25 +12,25 @@ use Doctrine\Common\Collections\ArrayCollection;
 class AvailableAttributes
 {
     /** @var array */
-    protected $attributeIds = [];
+    protected $attributes = [];
 
     /**
-     * Set attribute ids
+     * Set attribute
      *
-     * @param array $attributeIds
+     * @param array $attributes
      */
-    public function setAttributeIds(array $attributeIds)
+    public function setAttributes(array $attributes)
     {
-        $this->attributeIds = $attributeIds;
+        $this->attributes = $attributes;
     }
 
     /**
-     * Get attribute ids
+     * Get attribute
      *
      * @return array
      */
-    public function getAttributeIds()
+    public function getAttributes()
     {
-        return $this->attributeIds;
+        return $this->attributes;
     }
 }